feat:Created Schemas, models and userRouter.
Two schemas: userSchema and noteSchema were created. This were in turn used to create user and note models. Secondly, user route was designed and will function to return all registered users for the app. Finally, we userRouter was required in app.js, and used to get a specified response.

In addition, 2 packages were installed in the app: body-parser and cors (for authorized resource sharing with external third parties).