fix:changed the method used for patch endpoint from the combination of find() and updateOne() methods which didn't produce the desired results, to findOneAndUpdate() which works perfectly.
Please note that there is no issue with 404 error built to check if note(identified by its _id) belongs to the current user. Only ensure that the incorrect id for testing purpose has 24 characters. If not, this check is skipped and only caught in the catch block. This is then incorrectly marked as server error, code 500.